models separated for ease of use
