+++ tomcat_speech/models/multimodal_models.py
@@ -1,0 +1,335 @@
+# the models used for multimodal, multitask classification
+import sys
+
+import torch.nn as nn
+
+from tomcat_speech.models.audio_model_bases import AcousticOnlyForMultitask, IntermediateFusionAcousticOnlyModel
+from tomcat_speech.models.model_prediction_layers import PredictionLayer
+from tomcat_speech.models.multimodal_model_bases import IntermediateFusionMultimodalModel, EarlyFusionMultimodalModel, LateFusionMultimodalModel, MultimodalBaseDuplicateInput
+from tomcat_speech.models.text_model_bases import TextPlusPredictionLayer, IntermediateFusionTextOnlyModel
+
+
+class MultitaskModel(nn.Module):
+    """
+    A model combining base + output layers for multitask learning
+    """
+
+    def __init__(
+        self,
+        params,
+        num_embeddings=None,
+        pretrained_embeddings=None,
+        multi_dataset=True,
+        use_distilbert=False
+    ):
+        super(MultitaskModel, self).__init__()
+        # save whether there are multiple datasets
+        # if so, assumes each dataset has its own task
+        self.multi_dataset = multi_dataset
+
+        # let the network know if it's only using text features
+        self.text_only = params.text_only
+
+        # # set base of model
+        # comment this out and uncomment the below to try late fusion model
+        if params.audio_only is True:
+            # todo: update IntermediateFusionAcousticOnlyModel to work with distilbert
+            self.base = IntermediateFusionAcousticOnlyModel(
+                params, num_embeddings, pretrained_embeddings, use_distilbert
+            )
+        elif params.text_only is False:
+            if params.fusion_type.lower() == "early":
+                self.base = EarlyFusionMultimodalModel(
+                    params, num_embeddings, pretrained_embeddings, use_distilbert
+                )
+            elif params.fusion_type.lower() == "late":
+                self.base = LateFusionMultimodalModel(
+                    params, num_embeddings, pretrained_embeddings, use_distilbert
+                )
+            else:
+                self.base = IntermediateFusionMultimodalModel(
+                    params, num_embeddings, pretrained_embeddings, use_distilbert
+                )
+        else:
+            # todo: update IntermediateFusionTextOnlyModel to work with distilbert
+            self.base = IntermediateFusionTextOnlyModel(
+                params, num_embeddings, pretrained_embeddings, use_distilbert
+            )
+            # self.base = TextOnlyRNN(
+            #     params, num_embeddings, pretrained_embeddings
+            # )
+
+        # uncomment this and comment the above to try the late fusion model
+        # self.base = LateFusionMultimodalModel(
+        #     params, num_embeddings, pretrained_embeddings
+        # )
+
+        # set output layers
+        self.task_0_predictor = PredictionLayer(params, params.output_0_dim)
+        self.task_1_predictor = PredictionLayer(params, params.output_1_dim)
+        self.task_2_predictor = PredictionLayer(params, params.output_2_dim)
+        self.task_3_predictor = PredictionLayer(params, params.output_3_dim)
+        self.task_4_predictor = PredictionLayer(params, params.output_4_dim)
+
+    def forward(
+        self,
+        acoustic_input,
+        text_input,
+        speaker_input=None,
+        length_input=None,
+        acoustic_len_input=None,
+        gender_input=None,
+        task_num=0,
+        get_prob_dist=False,
+        return_penultimate_layer=False,
+        save_encoded_data=False
+    ):
+        # NOTE: if return_penultimate_layer, forward returns a (preds, penult) double
+
+        # call forward on base model
+        if self.text_only:
+            final_base_layer = self.base(
+                acoustic_input,
+                text_input,
+                speaker_input=speaker_input,
+                length_input=length_input,
+                gender_input=gender_input,
+            )
+        else:
+            final_base_layer = self.base(
+                acoustic_input,
+                text_input,
+                speaker_input=speaker_input,
+                length_input=length_input,
+                acoustic_len_input=acoustic_len_input,
+                gender_input=gender_input,
+            )
+
+        task_0_out = None
+        task_1_out = None
+        task_2_out = None
+        task_3_out = None
+        task_4_out = None
+
+        if not self.multi_dataset:
+            task_0_out = self.task_0_predictor(final_base_layer, get_prob_dist, return_penultimate_layer)
+            task_1_out = self.task_1_predictor(final_base_layer, get_prob_dist, return_penultimate_layer)
+            task_2_out = self.task_2_predictor(final_base_layer, get_prob_dist, return_penultimate_layer)
+            task_3_out = self.task_3_predictor(final_base_layer, get_prob_dist, return_penultimate_layer)
+            task_4_out = self.task_3_predictor(final_base_layer, get_prob_dist, return_penultimate_layer)
+        else:
+            if task_num == 0:
+                task_0_out = self.task_0_predictor(final_base_layer, get_prob_dist, return_penultimate_layer)
+            elif task_num == 1:
+                task_1_out = self.task_1_predictor(final_base_layer, get_prob_dist, return_penultimate_layer)
+            elif task_num == 2:
+                task_2_out = self.task_2_predictor(final_base_layer, get_prob_dist, return_penultimate_layer)
+            elif task_num == 3:
+                task_3_out = self.task_3_predictor(final_base_layer, get_prob_dist, return_penultimate_layer)
+            elif task_num == 4:
+                task_4_out = self.task_4_predictor(final_base_layer, get_prob_dist, return_penultimate_layer)
+            else:
+                sys.exit(f"Task {task_num} not defined")
+
+        if save_encoded_data:
+            return task_0_out, task_1_out, task_2_out, task_3_out, task_4_out, final_base_layer
+        else:
+            return task_0_out, task_1_out, task_2_out, task_3_out, task_4_out
+
+
+class MultitaskAcousticShared(nn.Module):
+    """
+    A model combining base + output layers for multitask learning
+    """
+
+    def __init__(
+        self,
+        params,
+        num_embeddings=None,
+        pretrained_embeddings=None,
+        multi_dataset=True,
+    ):
+        super(MultitaskAcousticShared, self).__init__()
+        # save whether there are multiple datasets
+        # if so, assumes each dataset has its own task
+        self.multi_dataset = multi_dataset
+
+        # # set base of model
+        # comment this out and uncomment the below to try late fusion model
+        self.acoustic_base = AcousticOnlyForMultitask(
+            params,
+            multi_dataset,
+            use_rnn=(not (params.avgd_acoustic or params.add_avging)),
+        )
+
+        # set output layers
+        self.task_0_predictor = TextPlusPredictionLayer(
+            params,
+            params.output_0_dim,
+            num_embeddings=num_embeddings,
+            pretrained_embeddings=pretrained_embeddings,
+            multi_dataset=multi_dataset,
+        )
+        self.task_1_predictor = TextPlusPredictionLayer(
+            params,
+            params.output_1_dim,
+            num_embeddings=num_embeddings,
+            pretrained_embeddings=pretrained_embeddings,
+            multi_dataset=multi_dataset,
+        )
+        self.task_2_predictor = TextPlusPredictionLayer(
+            params,
+            params.output_2_dim,
+            num_embeddings=num_embeddings,
+            pretrained_embeddings=pretrained_embeddings,
+            multi_dataset=multi_dataset,
+        )
+        self.task_3_predictor = TextPlusPredictionLayer(
+            params,
+            params.output_3_dim,
+            num_embeddings=num_embeddings,
+            pretrained_embeddings=pretrained_embeddings,
+            multi_dataset=multi_dataset,
+        )
+
+    def forward(
+        self,
+        acoustic_input,
+        text_input,
+        speaker_input=None,
+        length_input=None,
+        acoustic_len_input=None,
+        gender_input=None,
+        task_num=0,
+    ):
+        # call forward on base model
+        final_base_layer = self.acoustic_base(
+            acoustic_input, length_input=acoustic_len_input,
+        )
+
+        task_0_out = None
+        task_1_out = None
+        task_2_out = None
+        task_3_out = None
+
+        if not self.multi_dataset:
+            task_0_out = self.task_0_predictor(
+                final_base_layer, text_input, speaker_input, length_input, gender_input
+            )
+            task_1_out = self.task_1_predictor(
+                final_base_layer, text_input, speaker_input, length_input, gender_input
+            )
+            task_2_out = self.task_2_predictor(
+                final_base_layer, text_input, speaker_input, length_input, gender_input
+            )
+            task_3_out = self.task_3_predictor(
+                final_base_layer, text_input, speaker_input, length_input, gender_input
+            )
+        else:
+            if task_num == 0:
+                task_0_out = self.task_0_predictor(
+                    final_base_layer,
+                    text_input,
+                    speaker_input,
+                    length_input,
+                    gender_input,
+                )
+            elif task_num == 1:
+                task_1_out = self.task_1_predictor(
+                    final_base_layer,
+                    text_input,
+                    speaker_input,
+                    length_input,
+                    gender_input,
+                )
+            elif task_num == 2:
+                task_2_out = self.task_2_predictor(
+                    final_base_layer,
+                    text_input,
+                    speaker_input,
+                    length_input,
+                    gender_input,
+                )
+            elif task_num == 3:
+                task_3_out = self.task_3_predictor(
+                    final_base_layer,
+                    text_input,
+                    speaker_input,
+                    length_input,
+                    gender_input,
+                )
+            else:
+                sys.exit(f"Task {task_num} not defined")
+
+        return task_0_out, task_1_out, task_2_out, task_3_out
+
+
+class MultitaskDuplicateInputModel(nn.Module):
+    """
+    A multimodal model that duplicates input modality tensors
+    Inspired by:
+    Frustratingly easy domain adaptation
+    """
+
+    def __init__(
+        self,
+        params,
+        num_embeddings=None,
+        pretrained_embeddings=None,
+        multi_dataset=True,
+        num_tasks=2,
+    ):
+        super(MultitaskDuplicateInputModel, self).__init__()
+        # save whether there are multiple datasets
+        # if so, assumes each dataset has its own task
+        self.multi_dataset = multi_dataset
+
+        # # set base of model
+        # comment this out and uncomment the below to try late fusion model
+        self.base = MultimodalBaseDuplicateInput(
+            params, num_embeddings, pretrained_embeddings, num_tasks
+        )
+
+        # set output layers
+        self.task_0_predictor = PredictionLayer(params, params.output_0_dim)
+        self.task_1_predictor = PredictionLayer(params, params.output_1_dim)
+        self.task_2_predictor = PredictionLayer(params, params.output_2_dim)
+
+    def forward(
+        self,
+        acoustic_input,
+        text_input,
+        speaker_input=None,
+        length_input=None,
+        acoustic_len_input=None,
+        gender_input=None,
+        task_num=0,
+    ):
+        # call forward on base model
+        final_base_layer = self.base(
+            acoustic_input,
+            text_input,
+            length_input=length_input,
+            gender_input=gender_input,
+            task_num=task_num,
+        )
+
+        task_0_out = None
+        task_1_out = None
+        task_2_out = None
+
+        if not self.multi_dataset:
+            task_0_out = self.task_0_predictor(final_base_layer)
+            task_1_out = self.task_1_predictor(final_base_layer)
+            task_2_out = self.task_2_predictor(final_base_layer)
+        else:
+            if task_num == 0:
+                task_0_out = self.task_0_predictor(final_base_layer)
+            elif task_num == 1:
+                task_1_out = self.task_1_predictor(final_base_layer)
+            elif task_num == 2:
+                task_2_out = self.task_2_predictor(final_base_layer)
+            else:
+                sys.exit(f"Task {task_num} not defined")
+
+        return task_0_out, task_1_out, task_2_out
