+++ tomcat_speech/train_and_test_models/train_single_task.py
@@ -11,7 +11,7 @@ from datetime import date
 import random
 
 from tomcat_speech.models.train_and_test_models import train_and_predict, make_train_state
-from tomcat_speech.models.input_models import MultitaskModel
+from tomcat_speech.models.multimodal_models import MultitaskModel
 from tomcat_speech.models.plot_training import *
 from tomcat_speech.train_and_test_models.train_and_test_utils import set_cuda_and_seeds
